requirements.txt
arxiv==1.4.7
requests==2.31.0
schedule==1.2.0
openai==0.28.1
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.1
python-dotenv==1.0.0